Move sparse MLA wrapper into kernel package
